refactor(message): Move ip feature implementations to ip.rs and add a note about the condition under which to remove them
